Code Grad School (CGS) Day 1
Today begins a new chapter in my career development. I am going to enter into the next phase of intentional study of software craftsmanship. Much of that simply has to do with practice. Just as Doctors have residency, consider that is what this is. I need to write a lot of code each day in the full stack. It will help me learn what I need and want to for my own self and career.

Where other people have a blog, I have a lot of pages in a lot of places. MySpace is one of the pages that will help me keep me organized.

MySpace was my first fear that I missed the boat. I didn't miss the boat, but I did lose some time. I'm making up for lost time the only way one can, and that's by putting in time. I like Glitch, and I think it is what I thought MySpace _was_

Code Grad School (CGS) This is my endeavor to become an authentic Senior Software Engineer. The cost to become one is extremely high in time, concentration, and financial investment. Thankfully, I have all of those, and so do you. In the wake of Coronoavirus, I'm going to re-enroll in another round of self education about software development.

If the Internet doesn't last, then we have bigger problems on our hands. Like, what we're going to name our Pony. However, until further notice I believe the Internet is here to stay. And in doing so, I believe these content and hosting platforms will as well:

- AWS
- Blogger
- DigitalOcean
- Dropbox
- Flickr
- Glitch
- Github
- Heroku
- iCloud
- Soundcloud
- Spotify
- Tumblr
- Youtube

Those are the areas I will be adding content in one way or another. Notice there are no social media platforms on there. Social media must include an analytics and call to action component to it. If there is no content to created, then there is no point in promoting empty noise. I want to build a web of content across all the listed platforms before I begin the subject of promotion, which is a different skill than software engineering.